chore(wmg): move common
