+++ frontend/src/views/WheresMyGene/components/HeatMap/components/XAxisChart/index.tsx
@@ -5,12 +5,12 @@ import { EVENTS } from "src/common/analytics/events";
 import {
   DispatchContext,
   StateContext,
-} from "src/views/WheresMyGene/common/store";
+} from "src/views/WheresMyGeneV2/common/store";
 import {
   deleteSingleGene,
   selectGeneInfoFromXAxis,
   setXAxisHeight,
-} from "src/views/WheresMyGene/common/store/actions";
+} from "src/views/WheresMyGeneV2/common/store/actions";
 import { useDeleteGenes } from "../../hooks/useDeleteGenes";
 import { CHART_PADDING_PX, SELECTED_STYLE } from "../../style";
 import {
@@ -33,7 +33,7 @@ import InfoSVG from "src/common/images/info-sign-icon.svg";
 import {
   X_AXIS_CHART_HEIGHT_PX,
   X_AXIS_HOVER_CONTAINER_HEIGHT_PX,
-} from "src/views/WheresMyGene/common/constants";
+} from "src/views/WheresMyGeneV2/common/constants";
 import { EXCLUDE_IN_SCREENSHOT_CLASS_NAME } from "src/views/WheresMyGeneV2/components/GeneSearchBar/components/SaveExport";
 import GeneSearchBar from "src/views/WheresMyGeneV2/components/GeneSearchBar";
