+++ frontend/src/views/WheresMyGene/components/HeatMap/style.ts
@@ -8,7 +8,7 @@ import {
   CONTENT_WRAPPER_TOP_BOTTOM_PADDING_PX,
 } from "src/components/Layout/style";
 import { LEGEND_MARGIN_BOTTOM_PX } from "../../style";
-import { X_AXIS_CHART_HEIGHT_PX } from "../../common/constants";
+import { X_AXIS_CHART_HEIGHT_PX } from "../../../WheresMyGeneV2/common/constants";
 import { Autocomplete } from "@mui/material";
 import { TagFilter } from "@czi-sds/components";
 import { spacesS } from "src/common/theme";
