+++ frontend/src/views/WheresMyGeneV2/common/types.ts
@@ -1,7 +1,7 @@
 import { DefaultMenuSelectOption } from "@czi-sds/components";
 import { CSSProperties } from "react";
-import { ChartFormat } from "../../WheresMyGeneV2/components/HeatMap/components/Chart/types";
-import { CellTypeMetadata } from "../components/HeatMap/utils";
+import { ChartFormat } from "../components/HeatMap/components/Chart/types";
+import { CellTypeMetadata } from "../../WheresMyGene/components/HeatMap/utils";
 
 export interface Organism {
   id: string;
