+++ frontend/src/views/WheresMyGeneV2/components/Filters/components/Organism/connect.ts
@@ -1,11 +1,11 @@
 import { useContext, useEffect, useMemo } from "react";
-import { Organism as IOrganism } from "src/views/WheresMyGene/common/types";
+import { Organism as IOrganism } from "src/views/WheresMyGeneV2/common/types";
 import { useAvailableOrganisms } from "src/common/queries/wheresMyGene";
 import {
   DispatchContext,
   StateContext,
-} from "src/views/WheresMyGene/common/store";
-import { selectOrganism } from "src/views/WheresMyGene/common/store/actions";
+} from "src/views/WheresMyGeneV2/common/store";
+import { selectOrganism } from "src/views/WheresMyGeneV2/common/store/actions";
 import { track } from "src/common/analytics";
 import { EVENTS } from "src/common/analytics/events";
