+++ frontend/src/views/WheresMyGeneV2/components/Filters/components/Sort/constants.ts
@@ -1,5 +1,5 @@
 import { InputDropdownProps as IInputDropdownProps } from "@czi-sds/components";
-import { SORT_BY } from "src/views/WheresMyGene/common/types";
+import { SORT_BY } from "src/views/WheresMyGeneV2/common/types";
 
 export const DEFAULT_INPUT_DROPDOWN_PROPS: Partial<IInputDropdownProps> = {
   sdsStyle: "square",
