+++ frontend/src/views/WheresMyGeneV2/components/GeneSearchBar/components/SaveExport/csvUtils.ts
@@ -6,13 +6,13 @@ import {
 import {
   CompareId,
   getCompareOptionNameById,
-} from "src/views/WheresMyGene/common/constants";
-import { State } from "src/views/WheresMyGene/common/store";
+} from "src/views/WheresMyGeneV2/common/constants";
+import { State } from "src/views/WheresMyGeneV2/common/store";
 import { Props } from ".";
 import { generateAndCopyShareUrl } from "../ShareButton/utils";
 
 import { deserializeCellTypeMetadata } from "src/views/WheresMyGene/components/HeatMap/utils";
-import { ChartProps } from "src/views/WheresMyGene/common/types";
+import { ChartProps } from "src/views/WheresMyGeneV2/common/types";
 import { ChartFormat } from "src/views/WheresMyGeneV2/components/HeatMap/components/Chart/types";
 
 const NO_SELECTION_STRING = "No selection";
